phase 2/backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
